+++ internal/provider/common/docs.go
@@ -98,8 +98,8 @@ const STATUS_SPEC_REVISION_DESCRIPTION = "Spec revision"
 const STATUS_APPLIED_SPEC_REVISION_DESCRIPTION = "Applied spec revision"
 const STATUS_WAIT_FOR_APPLIED_SPEC_REVISION_DESCRIPTION = "Use this attribute to wait for the environment to be fully provisioned. It will long pull environment status until it matches the applied spec revision."
 const STATUS_PENDING_DELETE_DESCRIPTION = "`true` indicates that environment is pending deletion"
-const STATUS_LOAD_BALANCERS_DESCRIPTION = "Status of internal load balancer."
-const STATUS_LOAD_BALANCERS_INTERNAL_DESCRIPTION = "Status of load balancers."
+const STATUS_LOAD_BALANCERS_DESCRIPTION = "Load balancer status information."
+const STATUS_LOAD_BALANCERS_INTERNAL_DESCRIPTION = "Status of internal load balancer."
 const STATUS_LOAD_BALANCERS_ENDPOINT_SERVICE_NAME_DESCRIPTION = "VPC endpoint service name in $endpoint_service_id.$region.vpce.amazonaws.com format (if any)"
 
 // AWS descriptions.